Add HTML formatter middleware
